FHSS step support based on own UID

+++ src/src/FHSS.h
@@ -5,14 +5,14 @@
 #include <stdint.h>
 
 extern int_fast32_t volatile DRAM_ATTR FreqCorrection;
-#define FreqCorrectionMax  100000
-#define FreqCorrectionMin  -100000
+#define FreqCorrectionMax 100000
+#define FreqCorrectionMin -100000
 #define FreqCorrectionStep 61 //min freq step is ~ 61hz
 
 void ICACHE_RAM_ATTR FHSSresetFreqCorrection();
 
-void ICACHE_RAM_ATTR FHSSsetCurrIndex(uint8_t value);
-uint8_t ICACHE_RAM_ATTR FHSSgetCurrIndex();
+void ICACHE_RAM_ATTR FHSSsetCurrIndex(uint32_t value);
+uint32_t ICACHE_RAM_ATTR FHSSgetCurrIndex();
 void ICACHE_RAM_ATTR FHSSincCurrIndex();
 uint8_t ICACHE_RAM_ATTR FHSSgetCurrSequenceIndex();
